test(guide): move guide steps out of cmd dir
