+++ lib/cmds/guide.js
@@ -9,14 +9,14 @@ import { log, wrappedLog } from '../utils/log'
 import { asciiText } from '../utils/text'
 import { handleAsyncError as handle } from '../utils/async'
 
-import loginStep from './guide/step-login'
-import seedStep from './guide/step-seed'
-import createSpaceStep from './guide/step-create-space'
-import setupStep from './guide/step-setup'
-import devServerStep from './guide/step-dev-server'
-import finishStep from './guide/step-finish'
+import loginStep from '../guide/step-login'
+import seedStep from '../guide/step-seed'
+import createSpaceStep from '../guide/step-create-space'
+import setupStep from '../guide/step-setup'
+import devServerStep from '../guide/step-dev-server'
+import finishStep from '../guide/step-finish'
 
-import { AbortedError, GUIDE_MAX_WIDTH } from './guide/helpers'
+import { AbortedError, GUIDE_MAX_WIDTH } from '../guide/helpers'
 import { errorStyle, pathStyle } from '../utils/styles'
 
 inquirer.registerPrompt('directory', selectDirectory)
