+++ lib/guide/helpers.js
