+++ lib/guide/step-create-space.js
@@ -1,11 +1,11 @@
 import shellescape from 'shell-escape'
 
-import { spaceCreate } from '../space_cmds/create'
-import { log, wrappedLog } from '../../utils/log'
-import { separator } from '../../utils/text'
-import { highlightStyle, codeStyle } from '../../utils/styles'
-import { generateNumberEmoji } from '../../utils/emojis'
-import { confirmation } from '../../utils/actions'
+import { spaceCreate } from '../cmds/space_cmds/create'
+import { log, wrappedLog } from '../utils/log'
+import { separator } from '../utils/text'
+import { highlightStyle, codeStyle } from '../utils/styles'
+import { generateNumberEmoji } from '../utils/emojis'
+import { confirmation } from '../utils/actions'
 
 import { GUIDE_MAX_WIDTH, AbortedError } from './helpers'
