+++ lib/guide/step-dev-server.js
@@ -2,11 +2,11 @@ import execa from 'execa'
 import opn from 'opn'
 import treeKill from 'tree-kill'
 
-import { log, wrappedLog, logError } from '../../utils/log'
-import { separator } from '../../utils/text'
-import { highlightStyle, codeStyle } from '../../utils/styles'
-import { generateNumberEmoji } from '../../utils/emojis'
-import { confirmation } from '../../utils/actions'
+import { log, wrappedLog, logError } from '../utils/log'
+import { separator } from '../utils/text'
+import { highlightStyle, codeStyle } from '../utils/styles'
+import { generateNumberEmoji } from '../utils/emojis'
+import { confirmation } from '../utils/actions'
 
 import { GUIDE_MAX_WIDTH, AbortedError } from './helpers'
