+++ lib/guide/step-finish.js
@@ -1,15 +1,15 @@
 import { join, relative } from 'path'
 
 import chalk from 'chalk'
-import { readFile } from 'mz/fs'
+import { readFile } from 'fs'
 import markdown from 'markdown-cli'
 
-import { getContext } from '../../context'
-import { createManagementClient } from '../../utils/contentful-clients'
-import { log, success, wrappedLog } from '../../utils/log'
-import { frame, separator } from '../../utils/text'
-import { codeStyle, pathStyle } from '../../utils/styles'
-import { successEmoji } from '../../utils/emojis'
+import { getContext } from '../context'
+import { createManagementClient } from '../utils/contentful-clients'
+import { log, success, wrappedLog } from '../utils/log'
+import { frame, separator } from '../utils/text'
+import { codeStyle, pathStyle } from '../utils/styles'
+import { successEmoji } from '../utils/emojis'
 
 import { GUIDE_MAX_WIDTH } from './helpers'
