+++ lib/guide/step-login.js
@@ -1,9 +1,9 @@
-import { login } from '../login'
-import { getContext } from '../../context'
-import { log, wrappedLog } from '../../utils/log'
-import { separator } from '../../utils/text'
-import { highlightStyle, codeStyle } from '../../utils/styles'
-import { generateNumberEmoji } from '../../utils/emojis'
+import { login } from '../cmds/login'
+import { getContext } from '../context'
+import { log, wrappedLog } from '../utils/log'
+import { separator } from '../utils/text'
+import { highlightStyle, codeStyle } from '../utils/styles'
+import { generateNumberEmoji } from '../utils/emojis'
 
 import { GUIDE_MAX_WIDTH } from './helpers'
