+++ lib/guide/step-seed.js
@@ -1,11 +1,11 @@
 import shellescape from 'shell-escape'
 
-import { spaceSeed } from '../space_cmds/seed'
-import { log, wrappedLog } from '../../utils/log'
-import { separator } from '../../utils/text'
-import { highlightStyle, codeStyle } from '../../utils/styles'
-import { generateNumberEmoji } from '../../utils/emojis'
-import { confirmation } from '../../utils/actions'
+import { spaceSeed } from '../cmds/space_cmds/seed'
+import { log, wrappedLog } from '../utils/log'
+import { separator } from '../utils/text'
+import { highlightStyle, codeStyle } from '../utils/styles'
+import { generateNumberEmoji } from '../utils/emojis'
+import { confirmation } from '../utils/actions'
 
 import { GUIDE_MAX_WIDTH, AbortedError } from './helpers'
