+++ lib/guide/step-setup.js
@@ -5,13 +5,13 @@ import inquirer from 'inquirer'
 import Listr from 'listr'
 import execa from 'execa'
 
-import { accessTokenCreate } from '../space_cmds/accesstoken_cmds/create'
-import { getContext } from '../../context'
-import { getLatestGitHubRelease } from '../../utils/github'
-import { log, wrappedLog } from '../../utils/log'
-import { separator } from '../../utils/text'
-import { highlightStyle } from '../../utils/styles'
-import { generateNumberEmoji } from '../../utils/emojis'
+import { accessTokenCreate } from '../cmds/space_cmds/accesstoken_cmds/create'
+import { getContext } from '../context'
+import { getLatestGitHubRelease } from '../utils/github'
+import { log, wrappedLog } from '../utils/log'
+import { separator } from '../utils/text'
+import { highlightStyle } from '../utils/styles'
+import { generateNumberEmoji } from '../utils/emojis'
 
 import { GUIDE_MAX_WIDTH } from './helpers'
